Phase 2 - Fetch Data - Machine Learning - Deloy AWS/requirements.txt
flask
pandas
scipy
matplotlib
sklearn
pyrebase4

